+++ WorstUrlShortener/Share.iOS/ActionViewController.cs
@@ -6,6 +6,11 @@ using UIKit;
 using WorstUrlShortener.ViewModels;
 using WorstUrlShortener.Views;
 using Xamarin.Forms;
+using System.Diagnostics;
+
+/*
+ * TO Debug (breakpoints, output etc)you need to run this as the Startup project in VS
+ */
 
 namespace Action.iOS
 {
@@ -29,60 +34,42 @@ namespace Action.iOS
             base.ViewDidLoad();
 
             var inputItems = this.ExtensionContext.InputItems;
-            //var alert = UIAlertController.Create("TEST", $"Context {inputItems[0].AttributedContentText.Value}", UIAlertControllerStyle.Alert);
 
-            //var urlString = string.Empty;
-            var urlString = "ZIP, NADA";
+
+            global::Xamarin.Forms.Forms.Init();
+            var xfPage = new ShareExtensionPage();
+            var viewModel = (ExtensionViewModel)xfPage.BindingContext;
+            viewModel.FinishedCommand = new Command(() => DoneClicked(this));
+            var newController = xfPage.CreateViewController();
+            this.PresentModalViewController(newController, false);
+
+            var urlString = string.Empty;
             var inputItem = inputItems[0];
             var itemProvider = inputItem.Attachments?[0];
             if(itemProvider !=null && itemProvider.HasItemConformingTo("public.url"))
             {
-                var alert = UIAlertController.Create("TEST 1", $"WE HAVE URL {urlString}", UIAlertControllerStyle.Alert);
 
-                PresentViewController(alert, true, () =>
-                {
-                    itemProvider.LoadItem(UTType.URL,null, ((o, error) =>
+                    itemProvider.LoadItem(UTType.URL, null, ((o, error) =>
                     {
-                        /*if (o == null)
+                        if (o == null)
                         {
                             return;
-                        }*/
+                        }
 
                         var url = (NSUrl) o;
-                        urlString = url.ToString();
+                        urlString = url.AbsoluteString;
+                        // Debug.WriteLine("URL: " + urlString);
 
-                        var alert2 = UIAlertController.Create("TEST", $"Context {urlString}", UIAlertControllerStyle.Alert);
+                        viewModel.LongURL = urlString;
+                        // Debug.WriteLine("VM URL: " + viewModel.LongURL);
 
-                        PresentViewController(alert2, true, () =>
-                        {
-                            DispatchQueue.MainQueue.DispatchAfter(new DispatchTime(DispatchTime.Now, 5000000000), () =>
-                            {
-                                // Inform the host that we're done, so it un-blocks its UI. Note: Alternatively you could call super's -didSelectPost, which will similarly complete the extension context.
-                                ExtensionContext.CompleteRequest(new NSExtensionItem[0], null);
-                            });
-                        });
+                        // Can't present the XF View here, just times out, so present earlier and rely on the binding
+
                     }));
-
-                });
-
 
             }
 
-
-            /*
-            // Initialize Xamarin.Forms framework
-            global::Xamarin.Forms.Forms.Init();
-            // Create an instance of XF page with associated View Model
-            var xfPage = new ShareExtensionPage();
-            var viewModel = (ExtensionViewModel)xfPage.BindingContext;
-            //viewModel.Message = "Welcome to XF Page created from an iOS Extension";
-            // Override the behavior to complete the execution of the Extension when a user press the button
-            viewModel.FinishedCommand = new Command(() => DoneClicked(this));
-            // Convert XF page to a native UIViewController which can be consumed by the iOS Extension
-            var newController = xfPage.CreateViewController();
-            // Present new view controller as a regular view controller
-            this.PresentModalViewController(newController, false);
-            */
+            
         }
 
         partial void DoneClicked(NSObject sender)