+++ changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added ability to share URL to the system #34
+- Added iOS Share Extension support closes #35
 ### Changed
 - Moved shortening to viewmodel and auto copies url clipboard #36
 ### Fixed